refactor: create user funcionando
