--- src/infra/common/guards/test/access-by-tenant.guard.spec.ts
@@ -1,7 +1,0 @@
-import { AccessByTenantGuard } from './access-by-tenant.guard';
-
-describe('AccessByTenantGuard', () => {
-  it('should be defined', () => {
-    expect(new AccessByTenantGuard()).toBeDefined();
-  });
-});
