+++ src/infra/common/guards/test/permission.guard.spec.ts
@@ -2,6 +2,6 @@ import { PermissionGuard } from '../permission.guard';
 
 describe('PermissionGuard', () => {
   it.skip('should be defined', () => {
-    expect(new PermissionGuard()).toBeDefined();
+    // expect(new PermissionGuard()).toBeDefined();
   });
 });
