+++ src/modules/role/test/role.controller.spec.ts
@@ -1,6 +1,6 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { RoleController } from './role.controller';
-import { RoleService } from './role.service';
+import { RoleController } from '../role.controller';
+import { RoleService } from '../role.service';
 
 describe('RoleController', () => {
   let controller: RoleController;
